+++ src/api/mod.rs
@@ -111,10 +111,10 @@ fn encode_message(status: Result<String, Box<dyn Error>>, format : &str) -> Resu
     })
 }
 
-fn encode_station_url<A>(connection_new: &A, station: Option<StationItem>, ip: &str, format : &str, hours: u32) -> Result<rouille::Response, Box<dyn Error>> where A: DbConnection {
+fn encode_station_url<A>(connection_new: &A, station: Option<StationItem>, ip: &str, format : &str, seconds: u64) -> Result<rouille::Response, Box<dyn Error>> where A: DbConnection {
     Ok(match station {
         Some(station) => {
-            let _ = connection_new.increase_clicks(&ip, &station, hours);
+            let _ = connection_new.increase_clicks(&ip, &station, seconds);
             let station = station.into();
             match format {
                 "json" => {
@@ -208,7 +208,7 @@ fn encode_status(status: Status, format : &str, static_dir: &str) -> rouille::Re
     }
 }
 
-pub fn start<A: 'static +  std::clone::Clone>(connection_new: A, host : String, port : i32, threads : usize, server_name: &str, static_dir: &str, log_dir: &str, prometheus_exporter_enabled: bool, prometheus_exporter_prefix: &str, click_timeout_hours: u32) where A: DbConnection, A: std::marker::Send, A: std::marker::Sync {
+pub fn start<A: 'static +  std::clone::Clone>(connection_new: A, host : String, port : i32, threads : usize, server_name: &str, static_dir: &str, log_dir: &str, prometheus_exporter_enabled: bool, prometheus_exporter_prefix: &str, clicks_valid_timeout: u64) where A: DbConnection, A: std::marker::Send, A: std::marker::Sync {
     let listen_str = format!("{}:{}", host, port);
     info!("Listen on {} with {} threads", listen_str, threads);
     let x : Option<usize> = Some(threads);
@@ -217,7 +217,7 @@ pub fn start<A: 'static +  std::clone::Clone>(connection_new: A, host : String, 
     let log_dir = log_dir.to_string();
     let prometheus_exporter_prefix = prometheus_exporter_prefix.to_string();
     rouille::start_server_with_pool(listen_str, x, move |request| {
-        handle_connection(&connection_new, request, &y, &static_dir, &log_dir, prometheus_exporter_enabled, &prometheus_exporter_prefix, click_timeout_hours)
+        handle_connection(&connection_new, request, &y, &static_dir, &log_dir, prometheus_exporter_enabled, &prometheus_exporter_prefix, clicks_valid_timeout)
     });
 }
 
@@ -281,7 +281,7 @@ fn log_to_file(file_name: &str, line: &str) {
     }
 }
 
-fn handle_connection<A>(connection_new: &A, request: &rouille::Request, server_name: &str, static_dir: &str, log_dir: &str, prometheus_exporter_enabled: bool, prometheus_exporter_prefix: &str, click_timeout_hours: u32) -> rouille::Response where A: DbConnection {
+fn handle_connection<A>(connection_new: &A, request: &rouille::Request, server_name: &str, static_dir: &str, log_dir: &str, prometheus_exporter_enabled: bool, prometheus_exporter_prefix: &str, clicks_valid_timeout: u64) -> rouille::Response where A: DbConnection {
     let remote_ip: String = request.header("X-Forwarded-For").unwrap_or(&request.remote_addr().ip().to_string()).to_string();
     let referer: String = request.header("Referer").unwrap_or(&"-".to_string()).to_string();
     let user_agent: String = request.header("User-agent").unwrap_or(&"-".to_string()).to_string();
@@ -300,7 +300,7 @@ fn handle_connection<A>(connection_new: &A, request: &rouille::Request, server_n
         log_to_file(&log_file, &line);
     };
     rouille::log_custom(request, log_ok, log_err, || {
-        let result = handle_connection_internal(connection_new, request, server_name, static_dir, prometheus_exporter_enabled, prometheus_exporter_prefix, click_timeout_hours);
+        let result = handle_connection_internal(connection_new, request, server_name, static_dir, prometheus_exporter_enabled, prometheus_exporter_prefix, clicks_valid_timeout);
         match result {
             Ok(response) => response,
             Err(err) => rouille::Response::text(err.to_string()).with_status_code(500),
@@ -308,7 +308,7 @@ fn handle_connection<A>(connection_new: &A, request: &rouille::Request, server_n
     })
 }
 
-fn handle_connection_internal<A>(connection_new: &A, request: &rouille::Request, server_name: &str, static_dir: &str, prometheus_exporter_enabled: bool, prometheus_exporter_prefix: &str, click_timeout_hours: u32) -> Result<rouille::Response, Box<dyn std::error::Error>> where A: DbConnection {
+fn handle_connection_internal<A>(connection_new: &A, request: &rouille::Request, server_name: &str, static_dir: &str, prometheus_exporter_enabled: bool, prometheus_exporter_prefix: &str, clicks_valid_timeout: u64) -> Result<rouille::Response, Box<dyn std::error::Error>> where A: DbConnection {
     if request.method() != "POST" && request.method() != "GET" {
         return Ok(rouille::Response::empty_404());
     }
@@ -428,7 +428,7 @@ fn handle_connection_internal<A>(connection_new: &A, request: &rouille::Request,
             "tags" => Ok(add_cors(encode_extra(connection_new.get_extra("TagCache", "TagName", Some(String::from(parameter)), param_order, param_reverse, param_hidebroken)?, format, "tag")?)),
             "states" => Ok(add_cors(encode_states(connection_new.get_states(None, Some(String::from(parameter)), param_order, param_reverse, param_hidebroken)?, format)?)),
             "vote" => Ok(add_cors(encode_message(connection_new.vote_for_station(&remote_ip, get_only_first_item(connection_new.get_station_by_uuid(parameter)?)), format)?)),
-            "url" => Ok(add_cors(encode_station_url(connection_new, get_only_first_item(connection_new.get_station_by_uuid(parameter)?), &remote_ip, format, click_timeout_hours)?)),
+            "url" => Ok(add_cors(encode_station_url(connection_new, get_only_first_item(connection_new.get_station_by_uuid(parameter)?), &remote_ip, format, clicks_valid_timeout)?)),
             "stations" => {
                 match parameter {
                     "topvote" => Ok(add_cors(Station::get_response(connection_new.get_stations_topvote(999999)?.drain(..).map(|x| x.into()).collect(), format)?)),
@@ -457,7 +457,7 @@ fn handle_connection_internal<A>(connection_new: &A, request: &rouille::Request,
             let format = command;
             let command = parameter;
             match command {
-                "url" => Ok(add_cors(encode_station_url(connection_new, get_only_first_item(connection_new.get_station_by_uuid(search)?), &remote_ip, format, click_timeout_hours)?)),
+                "url" => Ok(add_cors(encode_station_url(connection_new, get_only_first_item(connection_new.get_station_by_uuid(search)?), &remote_ip, format, clicks_valid_timeout)?)),
                 _ => Ok(rouille::Response::empty_404()),
             }
         }else{