+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - CLEANUP: remove all non http/https content from favicon field of stations
 - ANSIBLE: disabled all apache2 logging by default
+- CONFIG: type errors do not get ignored with default value
+- CONFIG: use human readable durations instead of fixed seconds or hours
 
 ## [0.6.8] 2020-01-22
 ### Fixed